clean up

+++ src/main/java/net/aurora/firstmod/items/custom/ModSwordItem.java
@@ -47,7 +47,7 @@ public class ModSwordItem extends TieredItem {
     }
 
     public static ItemAttributeModifiers createAttributes(Tier p_330371_, float p_331976_, float p_332104_) {
-        return ItemAttributeModifiers.builder().add(Attributes.ATTACK_DAMAGE, new AttributeModifier(BASE_ATTACK_DAMAGE_ID, (double)(p_331976_ + p_330371_.getAttackDamageBonus()), Operation.ADD_VALUE), EquipmentSlotGroup.MAINHAND).add(Attributes.ATTACK_SPEED, new AttributeModifier(BASE_ATTACK_SPEED_ID, (double)p_332104_, Operation.ADD_VALUE), EquipmentSlotGroup.MAINHAND).build();
+        return ItemAttributeModifiers.builder().add(Attributes.ATTACK_DAMAGE, new AttributeModifier(BASE_ATTACK_DAMAGE_ID, p_331976_ + p_330371_.getAttackDamageBonus(), Operation.ADD_VALUE), EquipmentSlotGroup.MAINHAND).add(Attributes.ATTACK_SPEED, new AttributeModifier(BASE_ATTACK_SPEED_ID, p_332104_, Operation.ADD_VALUE), EquipmentSlotGroup.MAINHAND).build();
     }
 
     public boolean canAttackBlock(BlockState state, Level level, BlockPos pos, Player player) {
